AI Specialist Finder/requirements.txt
# Web Scraping
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
scrapy==2.11.0

# Social Media APIs
tweepy==4.14.0
linkedin-api==2.0.0

# Database
sqlalchemy==2.0.23

# Data Processing
pandas==2.1.3
numpy==1.25.2

# Natural Language Processing
nltk==3.8.1
textblob==0.17.1

# Geolocation
geopy==2.4.0
pycountry==22.3.5

# Configuration
python-dotenv==1.0.0

# Logging
loguru==0.7.2

# HTTP Client
httpx==0.25.2

# Testing Dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0
pytest-asyncio>=0.21.0
pytest-xdist>=3.0.0
psutil>=5.9.0
